+++ app/helpers/test_helpers.go
@@ -1,0 +1,6 @@
+package helpers
+
+// SimAppChainID hardcoded chainID for simulation
+const (
+	SimAppChainID = "gaia-app"
+)