+++ docs/.vuepress/components/home.vue
@@ -1,0 +1,11 @@
+<template>
+  <div></div>
+</template>
+
+<script>
+export default {
+  mounted() {
+    window.location.assign(`${this.$site.base}hub-overview/overview.html`)
+  }
+}
+</script>